+++ CHANGELOG.md
@@ -12,6 +12,15 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 
 * Deprecate `Dune::Precision` as the definition is not clear.
 
+* The deprecated class `LFEMatrix` has been removed.
+
+* The deprecated utility function `Impl::makeFunctionWithCallOperator()` has been removed.
+
+* The deprecated constructor of `LagrangeSimplexLocalFiniteElement` that was
+  templated wrt a `VertexMap` has been removed.
+
+* The deprecated class `PQkLocalFiniteElementCache` has been removed.
+
 # Release 2.10
 
 * Fixed LocalKey codim for the Crouzeix-Raviart element